1. add fraction and lang.big.num

+++ wheel/lang/time.go
@@ -2,22 +2,22 @@ package lang
 
 import "time"
 
-func GetNS(duration time.Duration) float64 {
-	return float64(duration / time.Nanosecond)
+func GetNS(duration time.Duration) int64 {
+	return int64(duration / time.Nanosecond)
 }
 
-func GetUS(duration time.Duration) float64 {
-	return float64(duration / time.Microsecond)
+func GetUS(duration time.Duration) int64 {
+	return int64(duration / time.Microsecond)
 }
 
-func GetMS(duration time.Duration) float64 {
-	return float64(duration / time.Millisecond)
+func GetMS(duration time.Duration) int64 {
+	return int64(duration / time.Millisecond)
 }
 
-func GetS(duration time.Duration) float64 {
-	return float64(duration / time.Second)
+func GetS(duration time.Duration) int64 {
+	return int64(duration / time.Second)
 }
 
-func GetMIN(duration time.Duration) float64 {
-	return float64(duration / time.Minute)
+func GetMIN(duration time.Duration) int64 {
+	return int64(duration / time.Minute)
 }
